+++ Leviasan.Sanlog.MSTest/NamedFormatStringUnitTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Globalization;
+using System.Text;
 
 namespace Leviasan.Sanlog.MSTest
 {
@@ -75,6 +76,50 @@ namespace Leviasan.Sanlog.MSTest
             Assert.AreEqual("4", namedFormatString.Segments[1].Format(CultureInfo.InvariantCulture, StringComparisonValue));
             Assert.AreEqual("2024-05-22T23:56:18", namedFormatString.Segments[2].Format(CultureInfo.InvariantCulture, DateTimeValue));
             Assert.AreEqual("Year month: 2024 May. StringComparison: 4. Sortable date/time: 2024-05-22T23:56:18.", namedFormatString.Format(CultureInfo.InvariantCulture, DateTimeValue, StringComparisonValue));
+        }
+        [TestMethod]
+        public void IndexComponentSingle()
+        {
+            var format = NamedFormatString.Parse("0x{0:X} {0:E} {0:N}");
+            Assert.AreEqual("0x7FFFFFFFFFFFFFFF 9.223372E+018 9,223,372,036,854,775,807.00", format.Format(CultureInfo.InvariantCulture, long.MaxValue));
+            
+            format = NamedFormatString.Parse("{DateTime:dddd MMMM}");
+            Assert.AreEqual(DateTime.Now.ToString("dddd MMMM", CultureInfo.InvariantCulture), format.Format(CultureInfo.InvariantCulture, DateTime.Now));
+        }
+        [TestMethod]
+        public void IndexComponentNotOrdered()
+        {
+            var format = "{1} {0} {2}";
+            var namedFormatString = NamedFormatString.Parse(format);
+            Assert.AreEqual("1", namedFormatString.Segments[0].Name);
+            Assert.AreEqual("0", namedFormatString.Segments[1].Name);
+            Assert.AreEqual("2", namedFormatString.Segments[2].Name);
+            Assert.AreEqual(1, namedFormatString.Segments[0].Index);
+            Assert.AreEqual(0, namedFormatString.Segments[1].Index);
+            Assert.AreEqual(2, namedFormatString.Segments[2].Index);
+            Assert.AreEqual(string.Format(CultureInfo.InvariantCulture, CompositeFormat.Parse(format), 0, 1, 2), namedFormatString.Format(CultureInfo.InvariantCulture, 0, 1, 2));
+        }
+        [TestMethod]
+        public void InvalidInputStringFormatException()
+        {
+            var formatException = Assert.ThrowsException<FormatException>(() => NamedFormatString.Parse("{1} {InvalidName} {2}"));
+            Assert.AreEqual("Input string was not in a correct format. Failure to parse near offset 5. Expected an ASCII digit.", formatException.Message);
+            formatException = Assert.ThrowsException<FormatException>(() => NamedFormatString.Parse("{InvalidName} {0} {2}"));
+            Assert.AreEqual("Input string was not in a correct format. Failure to parse near offset 15. Expected an ASCII digit.", formatException.Message);
+            formatException = Assert.ThrowsException<FormatException>(() => NamedFormatString.Parse("{1} {0} {InvalidName}"));
+            Assert.AreEqual("Input string was not in a correct format. Failure to parse near offset 9. Expected an ASCII digit.", formatException.Message);
+        }
+        [TestMethod]
+        public void AlignmentComponent()
+        {
+            var namedFormatString = NamedFormatString.Parse("{0,-20} {1,5:N1}");
+            Assert.AreEqual("Adam                  40.0", namedFormatString.Format(CultureInfo.InvariantCulture, "Adam", 40));
+        }
+        [TestMethod]
+        public void EscapingBraces()
+        {
+            var namedFormatString = NamedFormatString.Parse("{{{0:D}}}");
+            Assert.AreEqual("{6324}", namedFormatString.Format(CultureInfo.InvariantCulture, 6324));
         }
     }
 }